correction to video code
I am fairly certain the video shows code that does not work as planned.  He has the port mapping done incorrectly and I suspect it only works in this video because he happened to have an instance running in Docker that worked with the port (4001) that he used.  The only way I could get it to work was to set ports: in the docker-compose.yml = "- "XXXX:YYYY" where YYYY was referenced in index.js and XXXX was the localhost port in the browser request.